add 1 func in controller
